requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
python-dotenv
requests
pydantic
transformers
torch
torchaudio
librosa
soundfile
accelerate